requirements.txt
# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Video Processing
opencv-python>=4.8.0
moviepy>=1.0.3
yt-dlp>=2023.11.0

# AI/ML and NLP
openai>=1.3.0
tiktoken>=0.5.0

# Image and Diagram Generation
Pillow>=10.0.0
matplotlib>=3.8.0
graphviz>=0.20.1

# Data Processing
numpy>=1.24.0
pandas>=2.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.1

# Web Interface
Flask>=3.0.0
Werkzeug>=3.0.0

# Authentication and Database
Flask-SQLAlchemy>=3.1.1
Flask-JWT-Extended>=4.6.0
Flask-Bcrypt>=1.0.1
Flask-CORS>=4.0.0
Authlib>=1.3.0
cryptography>=41.0.0

# Database
SQLAlchemy>=2.0.23
alembic>=1.13.0
